Add constraints for protocol java generation (#3119)
* Add constraint validation for Protocol Java generation

* Add tests for Enum constraints

* Always Box primitives

* Fix boolean type

* Fix tests

* Add default to Protocol

* Fix validation

* Fix impacted tests

* Fix one validation edge case

* Flip test for constraints

+++ legend-engine-xts-mongodb/legend-engine-xt-nonrelationalStore-mongodb-executionPlan/src/main/java/org/finos/legend/engine/plan/execution/stores/mongodb/testsupport/MongoDBStoreTestConnectionFactory.java
@@ -57,7 +57,7 @@ public class MongoDBStoreTestConnectionFactory implements ConnectionFactoryExten
 
             MongoDBURL serverUrl = new MongoDBURL();
             serverUrl.baseUrl = inMemoryServer.baseUrl;
-            serverUrl.port = inMemoryServer.port;
+            serverUrl.port = (long)inMemoryServer.port;
             testDataSourceSpec.serverURLs = Lists.mutable.of(serverUrl);
             testConnection.dataSourceSpecification = testDataSourceSpec;
