+++ legend-engine-xts-mongodb/legend-engine-xt-nonrelationalStore-mongodb-grammar/src/main/java/org/finos/legend/engine/language/pure/grammar/from/deserializer/MongoDBQueryDeserializer.java
@@ -111,7 +111,7 @@ public class MongoDBQueryDeserializer extends StdDeserializer<DatabaseCommand>
             else if (jsonNode.isInt())
             {
                 IntTypeValue typeValue = new IntTypeValue();
-                typeValue.value = jsonNode.intValue();
+                typeValue.value = jsonNode.longValue();
                 literalValue.value = typeValue;
             }
             else if (jsonNode.isLong())
